// lib/types/jobsTypes.ts
export type JobsTypes = {
  id: number;
  position: string;
  salaryRange: string;
  workLocation: string;
  jobType: string;
  educationRequirement: string;
  departmentRequirement: string;
  experienceRequirement: string;
  languageRequirement: string;
  workNature: string;
  workHours: string[];
  skills: string[];
  leavePolicy: string;
  jobDescription: string;
  contactInfo: string;
  createDate: string;
};